Fix paywalls v2 header safe area handling
